feat: 任务列表写操作改动：只有项目成员才有权对任务列表做写操作；
任务列表写操作改动：只有项目成员才有权对任务列表做写操作；

+++ app/service/task_lists.js
@@ -51,8 +51,14 @@ class _objectName_Service extends Service {
   }
 
   async create(payload) {
-    const { ctx } = this;
+    const { ctx, service } = this;
     const { project_id } = payload;
+    // 非项目成员则无权创建此项目的任务列表
+    const project = await service.tasks.getProjectForMember(project_id);
+    if (!project) {
+      ctx.helper.body.UNAUTHORIZED({ ctx, msg: '非项目成员则无权创建此项目的任务列表' });
+      return false;
+    }
     const taskList = await ctx.model.TaskLists.findAll({
       where: {
         project_id,
@@ -64,15 +70,63 @@ class _objectName_Service extends Service {
   }
 
   async update(payload) {
-    const { ctx } = this;
+    const { ctx, service } = this;
+    const { project_id } = payload;
+    // 非项目成员则无权修改此项目的任务列表
+    const project = await service.tasks.getProjectForMember(project_id);
+    if (!project) {
+      ctx.helper.body.UNAUTHORIZED({ ctx, msg: '非项目成员则无权修改此项目的任务列表' });
+      return false;
+    }
     return await ctx.model.TaskLists.update(payload, {
       where: { id: payload.id },
       individualHooks: true,
     });
   }
 
+  /**
+   * 获取当前用户为项目成员的任务
+   * @param payload
+   * @return {Promise<*>}
+   */
+  async getTaskListForMember(payload) {
+    const { ctx } = this;
+    return await ctx.model.TaskLists.findAll({
+      where: {
+        id: payload,
+      },
+      include: [
+        {
+          model: ctx.model.Projects,
+          as: 'project',
+          required: true,
+          include: [
+            {
+              model: ctx.model.Users,
+              as: 'member',
+              where: {
+                id: ctx.currentRequestData.userInfo.id,
+              },
+            },
+          ],
+        },
+      ],
+    });
+  }
+
   async destroy(payload) {
     const { ctx } = this;
+    const taskListsExist = await ctx.model.TaskLists.findAll({
+      where: {
+        id: payload.ids,
+      },
+    });
+    // 如果存在任务列表数量和此用户为项目成员的任务列表数量不一致，则认为存在删除非项目成员任务列表，则 非项目成员则无权删除此项目的任务列表
+    const taskLists = await this.getTaskListForMember(payload.ids);
+    if (taskListsExist.length !== taskLists.length) {
+      ctx.helper.body.UNAUTHORIZED({ ctx, msg: '非项目成员则无权删除此项目的任务列表' });
+      return false;
+    }
     return await ctx.model.TaskLists.destroy({
       where: { id: payload.ids },
       individualHooks: true,
@@ -82,6 +136,12 @@ class _objectName_Service extends Service {
   async sort(payload) {
     const { ctx } = this;
     const { preId, nextId } = payload;
+    // 非项目成员则无权修改此项目的任务列表
+    const taskLists = await this.getTaskListForMember(payload.id);
+    if (!(taskLists && taskLists.length)) {
+      ctx.helper.body.UNAUTHORIZED({ ctx, msg: '非项目成员则无权修改此项目的任务列表' });
+      return false;
+    }
     let sort = 0;
     if (nextId !== undefined && preId !== undefined) {
       const pre = await ctx.model.TaskLists.findOne({ where: { id: preId } });
